requirements.txt
# ============================================
# DATA PROCESSING & UTILITIES
# ============================================
pandas>=1.3.5
python-dateutil>=2.8.2

# ============================================
# TESTING FRAMEWORK
# ============================================
pytest>=7.0.1
pytest-cov>=3.0.0
unittest-xml-reporting>=3.2.0

# ============================================
# DEVELOPMENT TOOLS
# ============================================
ipython>=8.0.1

# ============================================
# CODE QUALITY (Optional)
# ============================================
black>=22.1.0
pylint>=2.12.2

# ============================================
# RAG SYSTEM DEPENDENCIES
# ============================================
# Vector Database & Embeddings
lancedb>=0.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0

# LLM Integration
groq>=0.4.0

# Text Processing & Document Loading
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0

# Environment & Configuration
python-dotenv>=1.0.0

# Additional Dependencies
numpy>=1.24.0
pypdf>=3.0.0                    
tiktoken>=0.5.0                 

# ============================================
# AI-MODERNIZATION SYSTEM DEPENDENCIES
# ============================================
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Kafka for event-driven architecture
kafka-python==2.0.2

# File handling
aiofiles==23.2.1

# HTTP client
httpx==0.25.2

# Code analysis
astroid==3.0.1

# Testing
pytest-asyncio==0.21.1

# ============================================
# VS CODE EXTENSION DEPENDENCIES
# ============================================
# Note: VS Code extension uses Node.js (JavaScript)
# Python dependencies above are sufficient for the RAG system
# Node.js packages are managed in vscode-rag-extension/package.json

# RAG SYSTEM SETUP:
# - LanceDB: Free, local vector database (runs on your machine)
# - Sentence-Transformers: Free, local embeddings (no API needed)
# - Groq API: Fast inference (requires free API key from console.groq.com)
# - Configure API key in .env file: GROQ_API_KEY=your_key_here

# ============================================
# ACCOUNTS-MODERNIZATION DEPENDENCIES
# ============================================
# CLI and code analysis tools (already included above)
# Additional dependencies:
# - python-dotenv (already included)
# - astroid (already included)
# - groq (already included)
# - pytest (already included)
